pyproject.toml
[project]
name = "eventure"
version = "0.1.0"
description = "AI Event Coordination System for Student Organizations"
readme = "README.md"
authors = [
    { name = "Temirlan Tyulyubayev", email = "temikbjj@gmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.18.3",
    "bcrypt>=5.0.0",
    "email-validator>=2.3.0",
    "fastapi>=0.128.0",
    "psycopg2-binary>=2.9.11",
    "pydantic-settings>=2.12.0",
    "python-jose[cryptography]>=3.5.0",
    "sqlalchemy>=2.0.46",
    "uvicorn[standard]>=0.40.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.2",
]

[tool.uv]
package = false  

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["backend"]
